Add files via upload
Some of the code I wrote to generate visualizations for my poster, including rasters, heat-map PSTHs, and neural activity plots. Perhaps my favorite creation was rasters that are projected onto a continuous histogram of binned spike data, as shown in Figure 6 of my poster. (many thanks to the authors of MATLAB for Neuroscientists, a phenomenal book!)